add edit files functionality

+++ src/app/api/tickets/[slug]/route.ts
@@ -4,7 +4,8 @@ import { ObjectId } from "bson";
 import getErrorMessage from "@/utils/errorResponses";
 import { auth, isAdmin } from "@/auth";
 import { TicketEntity, ticketEntitySchema } from "../route";
-import { TCreateTicket } from "@/schemas/TicketSchema";
+import { TCreateTicket, TUpdateTicket } from "@/schemas/TicketSchema";
+import { deleteFileFromBucket, updateFiles } from "@/utils/bucketFiles";
 
 const DATABASE = process.env.MONGODB_DATABASE;
 const TICKETS_COLLECTION = "tickets";
@@ -23,7 +24,7 @@ export async function GET(
 
     if (!ticket) return NextResponse.json("ticket not found", { status: 404 });
     return NextResponse.json(
-      TCreateTicket.convertFromEntity({
+      TUpdateTicket.convertFromEntity({
         ...ticket,
       })
     );
@@ -43,17 +44,36 @@ export async function PUT(
     const session = await auth();
     if (!isAdmin(session))
       return NextResponse.json("Not enough privileges", { status: 401 });
-    const body = await req.json();
+
+    const prevTicket = await db
+      .collection<TicketEntity>(TICKETS_COLLECTION)
+      .findOne({ _id: new ObjectId(slug) });
+    if (!prevTicket) {
+      return NextResponse.json("Ticket does not exist", { status: 404 });
+    }
+
+    const formdata = await req.formData();
+
+    const files = formdata.getAll("files") as Array<File | string> | null;
+    const fileNamesAdded = await updateFiles(files, prevTicket.files);
+
     const updatedTicket = ticketEntitySchema
       .omit({ _id: true })
       .partial()
-      .parse(body);
+      .parse({
+        category: formdata.get("category"),
+        description: formdata.get("description"),
+        status: formdata.get("status"),
+        priority: formdata.get("priority"),
+        user_asigned: formdata.get("user_asigned"),
+        close_date: formdata.get("close_date"),
+        files: fileNamesAdded,
+      });
+
     const insertedUpdatedTicket = await db
       .collection<TicketEntity>(TICKETS_COLLECTION)
       .findOneAndUpdate(
-        {
-          _id: new ObjectId(slug),
-        },
+        { _id: new ObjectId(slug) },
         { $set: updatedTicket },
         { returnDocument: "after" }
       );
@@ -78,6 +98,17 @@ export async function DELETE(
     const session = await auth();
     if (!isAdmin(session))
       return NextResponse.json("Not enough privileges", { status: 401 });
+
+    const ticketToDelete = await db
+      .collection<TicketEntity>(TICKETS_COLLECTION)
+      .findOne({
+        _id: new ObjectId(slug),
+      });
+    if (!ticketToDelete)
+      return NextResponse.json("Ticket does not exist", { status: 404 });
+    ticketToDelete.files.map(async (file) => {
+      await deleteFileFromBucket(file);
+    });
     await db.collection<TicketEntity>(TICKETS_COLLECTION).deleteOne({
       _id: new ObjectId(slug),
     });
